--- a/cpp_asio_grpc_bench/main.cpp
+++ b/cpp_asio_grpc_bench/main.cpp
@@ -49,7 +49,7 @@ void spawn_accept_loop(
         return;
       }
       helloworld::HelloReply response;
-      response.set_message(context->request.name());
+      *response.mutable_response() = context->request.request();
       auto &writer = context->writer;
       agrpc::finish(writer, response, grpc::Status::OK,
                     boost::asio::bind_executor(