--- a/cpp_asio_grpc_bench/main.cpp
+++ b/cpp_asio_grpc_bench/main.cpp
@@ -49,7 +49,7 @@ void spawn_accept_loop(
         return;
       }
       helloworld::HelloReply response;
-      *response.mutable_response() = context->request.request();
+      *response.mutable_response() = std::move(*context->request.mutable_request());
       auto &writer = context->writer;
       agrpc::finish(writer, response, grpc::Status::OK,
                     boost::asio::bind_executor(