--- a/csharp_grpc_bench/GreeterServer/Program.cs
+++ b/csharp_grpc_bench/GreeterServer/Program.cs
@@ -24,7 +24,7 @@ namespace GreeterServer
         // Server side handler of the SayHello RPC
         public override Task<HelloReply> SayHello(HelloRequest request, ServerCallContext context)
         {
-            return Task.FromResult(new HelloReply { Message = request.Name });
+            return Task.FromResult(new HelloReply { Response = request.Request });
         }
     }