--- a/elixir_grpc_bench/lib/server.ex
+++ b/elixir_grpc_bench/lib/server.ex
@@ -4,6 +4,6 @@ defmodule Helloworld.Greeter.Server do
   @spec say_hello(Helloworld.HelloRequest.t(), GRPC.Server.Stream.t()) ::
           Helloworld.HelloReply.t()
   def say_hello(request, _stream) do
-    Helloworld.HelloReply.new(message: "#{request.name}")
+    Helloworld.HelloReply.new(response: request.request)
   end
 end