--- a/go_grpc_bench/example/main.go
+++ b/go_grpc_bench/example/main.go
@@ -39,7 +39,7 @@ type server struct {
 
 // SayHello implements helloworld.GreeterServer
 func (s *server) SayHello(ctx context.Context, in *pb.HelloRequest) (*pb.HelloReply, error) {
-	return &pb.HelloReply{Message: in.GetName()}, nil
+	return &pb.HelloReply{Response: in.GetRequest()}, nil
 }
 
 func main() {