--- a/java_quarkus_bench/src/main/java/com/example/HelloService.java
+++ b/java_quarkus_bench/src/main/java/com/example/HelloService.java
@@ -12,7 +12,7 @@ public class HelloService extends MutinyGreeterGrpc.GreeterImplBase {
     @Override
     public Uni<HelloReply> sayHello(HelloRequest request) {
         return Uni.createFrom().item(
-                HelloReply.newBuilder().setMessage(request.getName()).build()
+                HelloReply.newBuilder().setResponse(request.getRequest()).build()
         );
     }
 }