--- a/payload/100B
+++ b/payload/100B
@@ -1,1 +1,0 @@
-{"name":"The grpc server implementation is not as performant as we expected. This's extremely bad technology."}