--- a/php_swoole_grpc_onhold/Dockerfile
+++ b/php_swoole_grpc_onhold/Dockerfile
@@ -1,4 +1,4 @@
-FROM phpswoole/swoole:php8.1-alpine
+FROM phpswoole/swoole:php8.0-alpine
 
 WORKDIR /app