--- a/php_swoole_grpc_onhold/server.php
+++ b/php_swoole_grpc_onhold/server.php
@@ -8,7 +8,7 @@ require __DIR__ . '/vendor/autoload.php';
 
 $http = new Swoole\Http\Server('0.0.0.0', 50051, SWOOLE_BASE);
 $http->set([
-    'worker_num' => swoole_cpu_num(),
+    'trace_flags' => 0,
     'log_file' => '/dev/null',
     'log_level' => 5,
     'open_http2_protocol' => true
@@ -18,10 +18,9 @@ $http->on('request', function ($request, $response) {
 
     try {
         $request_message = new HelloRequest;
-        $request_message->mergeFromString(Parser::unpack($request->rawContent()));
 
         $response_message = new HelloReply();
-        $response_message->setMessage($request_message->getName());
+        $response_message->setResponse($request_message->getRequest());
 
         $response->header('content-type', 'application/grpc');
         $response->header('trailer', 'grpc-status, grpc-message');