--- a/python_async_grpc_bench/server.py
+++ b/python_async_grpc_bench/server.py
@@ -25,7 +25,7 @@ class Greeter(helloworld_pb2_grpc.GreeterServicer):
     async def SayHello(
             self, request: helloworld_pb2.HelloRequest,
             context: grpc.aio.ServicerContext) -> helloworld_pb2.HelloReply:
-        return helloworld_pb2.HelloReply(message=request.name)
+        return helloworld_pb2.HelloReply(response=request.request)
 
 
 async def serve() -> None: