--- a/python_grpc_bench/server.py
+++ b/python_grpc_bench/server.py
@@ -24,7 +24,7 @@ import helloworld_pb2_grpc
 class Greeter(helloworld_pb2_grpc.GreeterServicer):
 
     def SayHello(self, request, context):
-        return helloworld_pb2.HelloReply(message=request.name)
+        return helloworld_pb2.HelloReply(response=request.request)
 
 
 def serve():