--- a/rust_grpcio_bench/src/main.rs
+++ b/rust_grpcio_bench/src/main.rs
@@ -21,7 +21,7 @@ struct GreeterService;
 impl Greeter for GreeterService {
     fn say_hello(&mut self, ctx: RpcContext<'_>, req: HelloRequest, sink: UnarySink<HelloReply>) {
         let mut resp = HelloReply::default();
-        resp.set_message(req.get_name().to_string());
+        resp.set_response(req.get_request().clone());
         let f = sink
             .success(resp)
             .map_err(move |e| println!("failed to reply {:?}: {:?}", req, e))