--- a/rust_grpcio_bench/src/main.rs
+++ b/rust_grpcio_bench/src/main.rs
@@ -19,9 +19,9 @@ static GLOBAL: jemallocator::Jemalloc = jemallocator::Jemalloc;
 struct GreeterService;
 
 impl Greeter for GreeterService {
-    fn say_hello(&mut self, ctx: RpcContext<'_>, req: HelloRequest, sink: UnarySink<HelloReply>) {
+    fn say_hello(&mut self, ctx: RpcContext<'_>, mut req: HelloRequest, sink: UnarySink<HelloReply>) {
         let mut resp = HelloReply::default();
-        resp.set_response(req.get_request().clone());
+        resp.set_response(req.take_request());
         let f = sink
             .success(resp)
             .map_err(move |e| println!("failed to reply {:?}: {:?}", req, e))