--- a/scala_zio_bench/src/main/scala/com/example/helloworld/GreeterImpl.scala
+++ b/scala_zio_bench/src/main/scala/com/example/helloworld/GreeterImpl.scala
@@ -8,5 +8,5 @@ import zio.{ZEnv, ZIO}
 
 object GreeterImpl extends ZGreeter[ZEnv, Any] {
   override def sayHello(request: HelloRequest): ZIO[ZEnv, Status, HelloReply] =
-    ZIO.succeed(HelloReply(s"${request.name}"))
+    ZIO.succeed(HelloReply(request.request))
 }