--- a/scenarios/complex_proto/payload
+++ b/scenarios/complex_proto/payload
@@ -1,0 +1,1 @@
+{"request":{"name":"a name","d":4.55332,"f":232.3,"b":true,"n":32,"l":444325235223,"c1":"ofcouse","pets":[{"name":"Bof the dog","color":"BLUE"}, {"name":"Kim the cat","color":"RED"}]}}