--- a/scenarios/string_100B/payload
+++ b/scenarios/string_100B/payload
@@ -1,0 +1,1 @@
+{"request":{"name":"The grpc server implementation is not as performant as we expected. This's extremely bad technology."}}