--- a/setup_scenario.sh
+++ b/setup_scenario.sh
@@ -1,0 +1,14 @@
+#!/bin/sh
+
+SCENARIO=$1
+PAYLOAD=$2
+
+rm -rf proto
+mkdir -p proto/helloworld
+cp scenarios/"${SCENARIO}"/helloworld.proto proto/helloworld
+
+if ${COPY_PAYLOAD}; then
+	rm -rf payload
+	mkdir -p payload
+	cp scenarios/"${SCENARIO}"/payload payload/payload
+fi