… analyze.sh
… haskell_mu_onhold/.gitignore
… haskell_mu_onhold/Dockerfile
… haskell_mu_onhold/app/Main.hs
… haskell_mu_onhold/haskell-mu-bench.cabal
… haskell_mu_onhold/helloworld.proto
… haskell_mu_onhold/run.sh
… payload/10B
… payload/10kB
… payload/1kB
… php_swoole_grpc_onhold/composer.json
… rust_thruster_mt_bench/Cargo.lock
… rust_thruster_st_bench/Cargo.lock
… rust_tonic_mt_bench/Cargo.lock
… rust_tonic_st_bench/Cargo.lock
… scenarios/string_10B/payload
… scenarios/string_10kB/payload
… scenarios/string_1kB/payload
… swift_grpc_onhold/.dockerignore
… swift_grpc_onhold/Dockerfile
… swift_grpc_onhold/Package.swift
… swift_grpc_onhold/Sources/Model/helloworld.grpc.swift
… swift_grpc_onhold/Sources/Model/helloworld.pb.swift
… swift_grpc_onhold/Sources/Server/main.swift