The idea is to get rid of `GRPC_REQUEST_PAYLOAD` and replace it with a more general `GRPC_REQUEST_SCENARIO`.
Changing the scenario will make possible to customize not only the payload, but also the very protobuf contract between client and server. 
I'm encapsulating the old `string` field in a more general `message`. 
On the downside, changing scenario might require a rebuild of all images (if the protobuf changed); there isn't much we can do about this.

closes #169