@Trisfald 2022-01-08
- [ ] Adapt all benches to the new 'generic' proto