@LesnyRumcajs 2022-01-10
With this encapsulation, does the regular string scenario show the same / similar performance as before?