@Trisfald 2022-01-10
> With this encapsulation, does the regular string scenario show the same / similar performance as before?

I've run a subset of tests:

**Before:**
```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| rust_tonic_mt               |   89792 |        9.79 ms |       20.02 ms |       26.66 ms |       41.03 ms |  276.68% |     22.48 MiB |
| rust_thruster_mt            |   81971 |       10.84 ms |       21.75 ms |       28.85 ms |       44.47 ms |  273.32% |     16.12 MiB |
| rust_grpcio                 |   80275 |       11.15 ms |       17.19 ms |       20.46 ms |       28.85 ms |   285.1% |     34.21 MiB |
| java_hotspot_grpc_zgc       |   66493 |       12.78 ms |       28.96 ms |       51.18 ms |       65.54 ms |  305.11% |     127.3 MiB |
| java_hotspot_grpc_pgc       |   65062 |       12.78 ms |       31.72 ms |       50.96 ms |       65.83 ms |  304.05% |     191.8 MiB |
| java_hotspot_grpc_g1gc      |   61582 |       13.86 ms |       34.97 ms |       50.10 ms |       66.50 ms |  306.02% |     143.0 MiB |
| rust_tonic_st               |   55308 |       17.74 ms |       22.23 ms |       23.55 ms |       26.02 ms |  102.57% |     17.65 MiB |
| java_quarkus                |   52351 |       17.50 ms |       37.64 ms |       43.05 ms |       52.66 ms |  258.91% |    111.55 MiB |
| java_hotspot_grpc_sgc       |   50688 |       18.01 ms |       35.15 ms |       42.78 ms |       66.12 ms |  304.09% |     91.76 MiB |
| rust_thruster_st            |   48992 |       20.14 ms |       25.46 ms |       26.80 ms |       28.90 ms |   97.35% |      12.4 MiB |
| java_openj9_grpc_gencon     |   47801 |       18.18 ms |       43.59 ms |       60.21 ms |       86.71 ms |  301.94% |    108.41 MiB |
| java_hotspot_grpc_she       |   40951 |       22.37 ms |       58.31 ms |       66.18 ms |       76.33 ms |  301.99% |     382.9 MiB |
| java_micronaut_workstealing |   32804 |       29.17 ms |       66.53 ms |       75.93 ms |       88.29 ms |  308.76% |     302.4 MiB |
| java_micronaut              |   30325 |       31.73 ms |       70.82 ms |       76.64 ms |       88.96 ms |  307.76% |     305.0 MiB |
| java_aot                    |   13154 |       74.37 ms |      156.91 ms |      177.94 ms |      218.60 ms |  299.07% |    160.85 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

**After:**

```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| rust_tonic_mt               |   85779 |        9.87 ms |       20.07 ms |       26.36 ms |       39.50 ms |  270.94% |     21.59 MiB |
| rust_thruster_mt            |   79038 |       11.03 ms |       22.54 ms |       29.45 ms |       43.17 ms |  271.17% |     15.73 MiB |
| rust_grpcio                 |   78951 |       11.00 ms |       17.00 ms |       20.96 ms |       30.09 ms |  281.65% |     33.66 MiB |
| java_hotspot_grpc_zgc       |   62571 |       13.10 ms |       28.97 ms |       45.61 ms |       68.70 ms |  299.86% |    122.15 MiB |
| java_hotspot_grpc_pgc       |   61168 |       13.35 ms |       31.15 ms |       48.36 ms |       67.93 ms |  306.83% |    195.35 MiB |
| java_hotspot_grpc_g1gc      |   60643 |       13.87 ms |       35.78 ms |       49.24 ms |       64.31 ms |  303.83% |    155.65 MiB |
| rust_tonic_st               |   54433 |       18.02 ms |       22.52 ms |       23.90 ms |       26.58 ms |    99.7% |     16.09 MiB |
| java_quarkus                |   53582 |       16.78 ms |       35.71 ms |       42.42 ms |       50.71 ms |  253.98% |    111.95 MiB |
| java_hotspot_grpc_sgc       |   52328 |       17.02 ms |       32.68 ms |       38.98 ms |       58.08 ms |  294.02% |     93.93 MiB |
| java_openj9_grpc_gencon     |   49814 |       17.21 ms |       38.76 ms |       56.73 ms |       82.05 ms |  305.78% |    105.15 MiB |
| rust_thruster_st            |   48291 |       20.41 ms |       25.68 ms |       26.99 ms |       29.43 ms |   99.57% |     12.43 MiB |
| java_hotspot_grpc_she       |   46452 |       19.05 ms |       48.13 ms |       60.64 ms |       74.46 ms |  304.71% |    371.35 MiB |
| java_micronaut              |   32311 |       29.40 ms |       66.07 ms |       76.09 ms |       91.68 ms |  307.05% |     323.2 MiB |
| java_micronaut_workstealing |   26328 |       36.61 ms |       77.42 ms |       82.60 ms |       93.42 ms |  306.91% |     282.4 MiB |
| java_aot                    |   13240 |       73.78 ms |      152.01 ms |      175.22 ms |      210.32 ms |  281.19% |    182.15 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

The results are similar.

TBH I'm thinking to put the `complex_proto` scenario as the default.. it's more realistic.