@LesnyRumcajs 2022-01-10
> 
> The results are similar.

Great!

> 
> TBH I'm thinking to put the `complex_proto` scenario as the default.. it's more realistic.

I agree, it's more realistic for a proper protobuf use. And we'll still have the simple scenario for those that use gRPC to send a message with a single JSON string.