@LesnyRumcajs 2022-01-10
> I failed to update `php_swoole_grpc`, `swift_grpc` and `haskell_mu`.

Shall we create corresponding issues to re-enable them?